Feature Added
Create a thread that polls the gameScoreboard function for goal updates
in 60 second intervals. Display's notification to screen with team
names, score and game clock. Need fine tuned as far as when to launch
and turn off. For now if the Score Updates setting is set to true when
the user selects the Live game selection the thread starts and the only
way to shut it off is to open the settings menu and set Score Updates to
false.